Separated the API declaration and the network related configuration into different files
